requirements.txt
kafka-python
requests
psycopg2-binary
pymongo
redis           
python-dotenv
faker
